v2 implementation

+++ app/db/models/User.py
@@ -1,12 +1,15 @@
-from sqlalchemy import Column, BigInteger, Text, TIMESTAMP
+# app/db/models/user.py
+from sqlalchemy import Column, Integer, Text, TIMESTAMP
 from sqlalchemy.sql import func
-from .base import Base
+from app.db.models.base import Base
 
 class User(Base):
     __tablename__ = "users"
 
-    user_id = Column(BigInteger, primary_key=True)
+    user_id = Column(Integer, primary_key=True)
     username = Column(Text, nullable=False, unique=True)
     email = Column(Text, nullable=False, unique=True)
     password_hash = Column(Text, nullable=False)
-    created_at = Column(TIMESTAMP, server_default=func.now(), nullable=False)
+    created_at = Column(
+        TIMESTAMP, server_default=func.now(), nullable=False
+    )
